+++ src/main/java/service/BankAccountService.java
@@ -10,11 +10,11 @@ public interface BankAccountService {
 	void deleteAccount(int productId); // delete
 	List<BankAccountPojo> getAllBankAccounts();//read 
 	BankAccountPojo getAnAccount(int productId);//read
-	String addFunds(int funds, String accountType);// updates
-	String withdrawalFunds(int funds, String accountType);//updates
+	String addFunds(double funds, String accountType);// updates
+	String withdrawalFunds(double funds, String accountType);//updates
 	BankAccountPojo getId();//read
 	String login(String username, String password); //create
 	String checkFunds(String accountType);//read
 	String logout(String input);//update
-	String transferFunds(String fromAccount, int funds, String toAccount);//update
+	String transferFunds(String fromAccount, double funds, String toAccount);//update
 }
